requirements.txt
Pillow==8.3.0
aiogram==2.14.3
uvloop
ujson
cchardet
aiodns
aiohttp[speedups]